requirements.txt
Flask
celery
python-dotenv
SQLAlchemy
pywin32 ; platform_system == 'Windows'